Move Controller into main.swift
